+++ src/main/java/com/mas/demo/model/CleanerWithKeysRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface CleanerWithKeysRepository extends JpaRepository<CleanerWithKeys, Integer> {
+public interface CleanerWithKeysRepository extends JpaRepository<CleanerWithKeys, Long> {
 }
