+++ src/main/java/com/mas/demo/model/ReceptionistRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface ReceptionistRepository extends JpaRepository<Receptionist, Integer> {
+public interface ReceptionistRepository extends JpaRepository<Receptionist, Long> {
 }
