+++ src/main/java/com/mas/demo/model/TraineeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface TraineeRepository extends JpaRepository<Room, Integer> {
+public interface TraineeRepository extends JpaRepository<Room, Long> {
 }
