+++ src/main/java/com/mas/demo/model/WorkshopsRepository.java
@@ -2,5 +2,5 @@ package com.mas.demo.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkshopsRepository extends JpaRepository<Workshops, Integer> {
+public interface WorkshopsRepository extends JpaRepository<Workshops, Long> {
 }
